Move metrics utils
