+++ hazelcast-jet-core/src/test/java/com/hazelcast/jet/core/JobRestartWithSnapshotTest.java
@@ -283,7 +283,9 @@ public class JobRestartWithSnapshotTest extends JetTestSupport {
 
         // instance1 is always coordinator
         delayOperationsFrom(
-                hz(instance1), JetInitDataSerializerHook.FACTORY_ID, singletonList(JetInitDataSerializerHook.SNAPSHOT_OPERATION)
+                hz(instance1),
+                JetInitDataSerializerHook.FACTORY_ID,
+                singletonList(JetInitDataSerializerHook.SNAPSHOT_OPERATION)
         );
 
         DAG dag = new DAG();
